requirements.txt
dj-config-url==0.1.1
dj-database-url==0.4.1
Django==1.8.13
django-crispy-forms==1.6.1
djangorestframework==3.5.3
gunicorn==19.6.0
psycopg2==2.6.2
reverse==0.1.0
whitenoise==3.2.2
